fix: bring all changed files under 200-line limit
- AC.integration.test.ts: 527→59 lines (full suite in AC-US1.3-*.test.ts + AC-US1.5.integration.test.ts)
- useBedFilter.test.ts: 239→133 lines (initial state / filter / sort moved to useBedFilter-state.test.ts)
- BedCard.tsx: 205→187 lines (highlightMatch extracted to lib/highlight-match.tsx)
- New: lib/highlight-match.tsx (21 lines)

All checks: 0 TS errors, 283/283 tests pass
