+++ src/features/bed-dashboard/__tests__/AC-US1.3-time-advanced.test.ts
@@ -1,0 +1,101 @@
+import { describe, it, expect } from 'vitest'
+
+// US-1.3 — Advanced integration: delay detection, edge cases, performance
+describe('US-1.3 Acceptance Criteria — Advanced Integration', () => {
+  describe('Delay Detection Integration', () => {
+    it('should identify delayed patients (> 3 hours)', () => {
+      const DELAY_THRESHOLD_MS = 10800000 // 3 hours
+      const elapsedTime = 11700000 // 3h 15m
+
+      const isDelayed = elapsedTime > DELAY_THRESHOLD_MS
+      expect(isDelayed).toBe(true)
+    })
+
+    it('should identify on-time patients (< 3 hours)', () => {
+      const DELAY_THRESHOLD_MS = 10800000
+      const elapsedTime = 9900000 // 2h 45m
+
+      const isDelayed = elapsedTime > DELAY_THRESHOLD_MS
+      expect(isDelayed).toBe(false)
+    })
+
+    it('should apply red color to delayed patients', () => {
+      const isDelayed = true
+      const textColor = isDelayed ? 'text-red-400' : 'text-zinc-300'
+      expect(textColor).toBe('text-red-400')
+    })
+
+    it('should apply normal color to on-time patients', () => {
+      const isDelayed = false
+      const textColor = isDelayed ? 'text-red-400' : 'text-zinc-300'
+      expect(textColor).toBe('text-zinc-300')
+    })
+  })
+
+  describe('Edge Cases', () => {
+    it('should handle patient just admitted', () => {
+      const admissionTime = Date.now() - 10000 // 10 seconds ago
+      const elapsed = Date.now() - admissionTime
+
+      const totalMinutes = Math.floor(elapsed / 60000)
+      const formatted = totalMinutes < 1 ? '< 1m' : `${totalMinutes}m`
+
+      expect(formatted).toBe('< 1m')
+    })
+
+    it('should handle multi-day patients', () => {
+      const admissionTime = Date.now() - 172800000 // 48 hours ago
+      const elapsed = Date.now() - admissionTime
+
+      const totalMinutes = Math.floor(elapsed / 60000)
+      const hours = Math.floor(totalMinutes / 60)
+      const minutes = totalMinutes % 60
+
+      expect(hours).toBe(48)
+      expect(minutes).toBeLessThan(60)
+    })
+
+    it('should handle null/invalid admission time', () => {
+      const admissionTime = null
+      const formatted = admissionTime ? 'valid' : 'N/A'
+      expect(formatted).toBe('N/A')
+    })
+  })
+
+  describe('Performance Tests', () => {
+    it('should calculate format in < 1ms', () => {
+      const formatElapsedTime = (ms: number) => {
+        const totalMinutes = Math.floor(ms / 60000)
+        if (totalMinutes < 1) return '< 1m'
+        const hours = Math.floor(totalMinutes / 60)
+        const minutes = totalMinutes % 60
+        if (hours === 0) return `${minutes}m`
+        return minutes === 0 ? `${hours}h` : `${hours}h ${minutes}m`
+      }
+
+      const start = performance.now()
+      formatElapsedTime(9900000)
+      const end = performance.now()
+
+      expect(end - start).toBeLessThan(1)
+    })
+
+    it('should handle 12 beds updating every minute efficiently', () => {
+      const bedCount = 12
+      const updateCount = 60 // 1 hour of updates
+
+      const start = performance.now()
+      for (let i = 0; i < bedCount * updateCount; i++) {
+        const elapsed = 9900000
+        const totalMinutes = Math.floor(elapsed / 60000)
+        const hours = Math.floor(totalMinutes / 60)
+        const mins = totalMinutes % 60
+        const formatted = `${hours}h ${mins}m`
+        expect(formatted).toBeDefined()
+      }
+      const end = performance.now()
+
+      expect(end - start).toBeLessThan(100)
+    })
+  })
+})
