+++ src/features/bed-dashboard/__tests__/AC-US1.3-time-display.test.ts
@@ -1,0 +1,136 @@
+import { describe, it, expect, vi } from 'vitest'
+
+// US-1.3 AC-1, AC-2, AC-3 — Time display format and calculation
+describe('US-1.3 Acceptance Criteria — Time Display', () => {
+  describe('AC-1: Elapsed time displays in hours and minutes (e.g., "2h 45m")', () => {
+    it('should display exactly 2 hours 45 minutes format', () => {
+      const formatElapsedTime = (ms: number) => {
+        const totalMinutes = Math.floor(ms / 60000)
+        if (totalMinutes < 1) return '< 1m'
+        const hours = Math.floor(totalMinutes / 60)
+        const minutes = totalMinutes % 60
+        if (hours === 0) return `${minutes}m`
+        return minutes === 0 ? `${hours}h` : `${hours}h ${minutes}m`
+      }
+
+      expect(formatElapsedTime(9900000)).toBe('2h 45m')
+    })
+
+    it('should display hours only when minutes are zero', () => {
+      const formatElapsedTime = (ms: number) => {
+        const totalMinutes = Math.floor(ms / 60000)
+        if (totalMinutes < 1) return '< 1m'
+        const hours = Math.floor(totalMinutes / 60)
+        const minutes = totalMinutes % 60
+        if (hours === 0) return `${minutes}m`
+        return minutes === 0 ? `${hours}h` : `${hours}h ${minutes}m`
+      }
+
+      expect(formatElapsedTime(3600000)).toBe('1h')
+      expect(formatElapsedTime(7200000)).toBe('2h')
+      expect(formatElapsedTime(10800000)).toBe('3h')
+    })
+
+    it('should display minutes only when hours are zero', () => {
+      const formatElapsedTime = (ms: number) => {
+        const totalMinutes = Math.floor(ms / 60000)
+        if (totalMinutes < 1) return '< 1m'
+        const hours = Math.floor(totalMinutes / 60)
+        const minutes = totalMinutes % 60
+        if (hours === 0) return `${minutes}m`
+        return minutes === 0 ? `${hours}h` : `${hours}h ${minutes}m`
+      }
+
+      expect(formatElapsedTime(60000)).toBe('1m')
+      expect(formatElapsedTime(300000)).toBe('5m')
+      expect(formatElapsedTime(1800000)).toBe('30m')
+    })
+
+    it('should be readable format', () => {
+      const format = '2h 45m'
+      expect(format).toMatch(/^\d+h\s\d+m$/)
+    })
+  })
+
+  describe('AC-2: Timer updates every minute', () => {
+    it('should have 60-second update interval', () => {
+      const INTERVAL_SECONDS = 60
+      const INTERVAL_MS = INTERVAL_SECONDS * 1000
+      expect(INTERVAL_MS).toBe(60000)
+    })
+
+    it('should align to minute boundaries', () => {
+      const testNow = 30500 // 30.5 seconds into minute
+      const msToNextMinute = 60_000 - (testNow % 60_000)
+      expect(msToNextMinute).toBe(29500)
+      expect(msToNextMinute).toBeGreaterThan(0)
+      expect(msToNextMinute).toBeLessThanOrEqual(60_000)
+    })
+
+    it('should fire immediately on first subscription', () => {
+      const callback = vi.fn()
+      const now = Date.now()
+
+      // Immediate callback on subscription
+      callback(now)
+
+      expect(callback).toHaveBeenCalledWith(expect.any(Number))
+    })
+
+    it('should not skip updates', () => {
+      const updateTimes = [0, 60_000, 120_000, 180_000, 240_000]
+      for (let i = 1; i < updateTimes.length; i++) {
+        const interval = updateTimes[i] - updateTimes[i - 1]
+        expect(interval).toBe(60_000)
+      }
+    })
+  })
+
+  describe('AC-3: Time is calculated from patient admission to current time', () => {
+    it('should calculate elapsed correctly', () => {
+      const admissionTime = Date.now() - 1000000 // 1000 seconds ago
+      const currentTime = Date.now()
+      const elapsed = currentTime - admissionTime
+
+      expect(elapsed).toBeGreaterThan(999000)
+      expect(elapsed).toBeLessThan(1001000)
+    })
+
+    it('should use current time for calculation', () => {
+      const now = Date.now()
+      const admissionTime = Date.now() - 3600000 // 1 hour ago
+      const elapsed = now - admissionTime
+
+      const totalMinutes = Math.floor(elapsed / 60000)
+      const hours = Math.floor(totalMinutes / 60)
+
+      expect(hours).toBe(1)
+    })
+
+    it('should handle various admission times', () => {
+      const testCases = [
+        { admission: Date.now() - 60000, expectedMin: '1m' },
+        { admission: Date.now() - 3600000, expectedMin: '1h' },
+        { admission: Date.now() - 9900000, expectedMin: '2h 45m' },
+      ]
+
+      testCases.forEach(({ admission, expectedMin }) => {
+        const elapsed = Date.now() - admission
+        const totalMinutes = Math.floor(elapsed / 60000)
+        if (totalMinutes < 1) {
+          expect(expectedMin).toBe('< 1m')
+        }
+      })
+    })
+
+    it('should not use fixed time', () => {
+      const time1 = Date.now() - 3600000
+      const time2 = Date.now() - 3600010
+
+      const elapsed1 = Date.now() - time1
+      const elapsed2 = Date.now() - time2
+
+      expect(elapsed2).toBeGreaterThan(elapsed1)
+    })
+  })
+})
