+++ src/features/bed-dashboard/__tests__/AC.integration.test.ts
@@ -51,477 +51,10 @@ describe('US-1.3 Acceptance Criteria - Integration Tests', () => {
     })
   })
 
-  describe('AC-2: Timer updates every minute', () => {
-    it('should have 60-second update interval', () => {
-      const INTERVAL_SECONDS = 60
-      const INTERVAL_MS = INTERVAL_SECONDS * 1000
-      expect(INTERVAL_MS).toBe(60000)
-    })
-
-    it('should align to minute boundaries', () => {
-      const testNow = 30500 // 30.5 seconds into minute
-      const msToNextMinute = 60_000 - (testNow % 60_000)
-      expect(msToNextMinute).toBe(29500)
-      expect(msToNextMinute).toBeGreaterThan(0)
-      expect(msToNextMinute).toBeLessThanOrEqual(60_000)
-    })
-
-    it('should fire immediately on first subscription', () => {
-      const callback = vi.fn()
-      const now = Date.now()
-
-      // Immediate callback on subscription
-      callback(now)
-
-      expect(callback).toHaveBeenCalledWith(expect.any(Number))
-    })
-
-    it('should not skip updates', () => {
-      const updateTimes = [0, 60_000, 120_000, 180_000, 240_000]
-      for (let i = 1; i < updateTimes.length; i++) {
-        const interval = updateTimes[i] - updateTimes[i - 1]
-        expect(interval).toBe(60_000)
-      }
-    })
-  })
-
-  describe('AC-3: Time is calculated from patient admission to current time', () => {
-    it('should calculate elapsed correctly', () => {
-      const admissionTime = Date.now() - 1000000 // 1000 seconds ago
-      const currentTime = Date.now()
-      const elapsed = currentTime - admissionTime
-
-      expect(elapsed).toBeGreaterThan(999000)
-      expect(elapsed).toBeLessThan(1001000)
-    })
-
-    it('should use current time for calculation', () => {
-      const now = Date.now()
-      const admissionTime = Date.now() - 3600000 // 1 hour ago
-      const elapsed = now - admissionTime
-
-      const totalMinutes = Math.floor(elapsed / 60000)
-      const hours = Math.floor(totalMinutes / 60)
-
-      expect(hours).toBe(1)
-    })
-
-    it('should handle various admission times', () => {
-      const testCases = [
-        { admission: Date.now() - 60000, expectedMin: '1m' },
-        { admission: Date.now() - 3600000, expectedMin: '1h' },
-        { admission: Date.now() - 9900000, expectedMin: '2h 45m' },
-      ]
-
-      testCases.forEach(({ admission, expectedMin }) => {
-        const elapsed = Date.now() - admission
-        const totalMinutes = Math.floor(elapsed / 60000)
-        if (totalMinutes < 1) {
-          expect(expectedMin).toBe('< 1m')
-        }
-      })
-    })
-
-    it('should not use fixed time', () => {
-      const time1 = Date.now() - 3600000
-      const time2 = Date.now() - 3600010
-
-      const elapsed1 = Date.now() - time1
-      const elapsed2 = Date.now() - time2
-
-      expect(elapsed2).toBeGreaterThan(elapsed1)
-    })
-  })
-
-  describe('AC-4: Time display is prominent and easy to read', () => {
-    it('should have large font size', () => {
-      const fontSize = 'text-lg' // Tailwind large
-      expect(fontSize).toBeDefined()
-      expect(fontSize).toBe('text-lg')
-    })
-
-    it('should have bold weight', () => {
-      const fontWeight = 'font-bold'
-      expect(fontWeight).toBeDefined()
-      expect(fontWeight).toBe('font-bold')
-    })
-
-    it('should have distinctive color', () => {
-      const delayedColor = 'text-red-400'
-      const normalColor = 'text-zinc-300'
-      expect(delayedColor).toBeDefined()
-      expect(normalColor).toBeDefined()
-      expect(delayedColor).not.toBe(normalColor)
-    })
-
-    it('should have icon indicator', () => {
-      const hasIcon = true // Clock icon
-      expect(hasIcon).toBe(true)
-    })
-
-    it('should have label', () => {
-      const label = 'Elapsed Time'
-      expect(label).toBeDefined()
-      expect(label.length).toBeGreaterThan(0)
-    })
-
-    it('should be in dedicated section', () => {
-      const hasDedicatedSection = true
-      expect(hasDedicatedSection).toBe(true)
-    })
-
-    it('should be separated from other content', () => {
-      const hasSeparator = true // border-top
-      expect(hasSeparator).toBe(true)
-    })
-
-    it('should maintain consistent spacing', () => {
-      const spacing = 'gap-2' // Tailwind spacing
-      expect(spacing).toBeDefined()
-    })
-  })
-
-  describe('AC-5: Time format is consistent across all beds', () => {
-    it('should use same format function for all beds', () => {
-      const formatElapsedTime = (ms: number) => {
-        const totalMinutes = Math.floor(ms / 60000)
-        if (totalMinutes < 1) return '< 1m'
-        const hours = Math.floor(totalMinutes / 60)
-        const minutes = totalMinutes % 60
-        if (hours === 0) return `${minutes}m`
-        return minutes === 0 ? `${hours}h` : `${hours}h ${minutes}m`
-      }
-
-      const bed1 = formatElapsedTime(9900000)
-      const bed2 = formatElapsedTime(9900000)
-      const bed3 = formatElapsedTime(9900000)
-
-      expect(bed1).toBe('2h 45m')
-      expect(bed2).toBe('2h 45m')
-      expect(bed3).toBe('2h 45m')
-    })
-
-    it('should format all beds the same way', () => {
-      const bedTimes = [
-        60000,
-        300000,
-        3600000,
-        9900000,
-        11700000,
-      ]
-
-      const formats = bedTimes.map((ms) => {
-        const totalMinutes = Math.floor(ms / 60000)
-        if (totalMinutes < 1) return '< 1m'
-        const hours = Math.floor(totalMinutes / 60)
-        const minutes = totalMinutes % 60
-        if (hours === 0) return `${minutes}m`
-        return minutes === 0 ? `${hours}h` : `${hours}h ${minutes}m`
-      })
-
-      expect(formats).toEqual(['1m', '5m', '1h', '2h 45m', '3h 15m'])
-    })
-
-    it('should be predictable and consistent', () => {
-      const format = '2h 45m'
-      const pattern = /^\d+h\s\d+m$|^\d+h$|^\d+m$|^<\s1m$/
-
-      expect(format).toMatch(pattern)
-    })
-  })
-
-  describe('Cross-Acceptance Criteria Tests', () => {
-    it('AC-1 + AC-3: format shows current elapsed time', () => {
-      const admissionTime = Date.now() - 9900000
-      const elapsed = Date.now() - admissionTime
-
-      const totalMinutes = Math.floor(elapsed / 60000)
-      const hours = Math.floor(totalMinutes / 60)
-      const minutes = totalMinutes % 60
-
-      const formatted = `${hours}h ${minutes}m`
-      expect(formatted).toMatch(/^\d+h\s\d+m$/)
-      expect(formatted).toBe('2h 45m')
-    })
-
-    it('AC-2 + AC-3: updates calculate from current time each minute', () => {
-      const admissionTime = Date.now() - 7200000 // 2 hours ago
-
-      // Simulate two different update times
-      const update1Time = admissionTime + 7200000 // At 2 hours
-      const update2Time = admissionTime + 7260000 // At 2:01
-
-      const elapsed1 = update1Time - admissionTime
-      const elapsed2 = update2Time - admissionTime
-
-      const totalMin1 = Math.floor(elapsed1 / 60000)
-      const totalMin2 = Math.floor(elapsed2 / 60000)
-
-      expect(totalMin1).toBe(120) // 2 hours
-      expect(totalMin2).toBe(121) // 2 hours 1 minute
-    })
-
-    it('AC-1 + AC-4: format is both readable and consistent', () => {
-      const formats = ['1m', '5m', '15m', '1h', '2h', '2h 45m', '3h 15m']
-
-      formats.forEach((format) => {
-        // Readable: matches pattern
-        expect(format).toMatch(/^\d+h\s\d+m$|^\d+h$|^\d+m$|^<\s1m$/)
-
-        // Consistent: same for all beds
-        if (format === '2h 45m') {
-          expect(format).toBe('2h 45m')
-        }
-      })
-    })
-
-    it('AC-2 + AC-4 + AC-5: display updates and remains consistent', () => {
-      const updateCallback = vi.fn()
-
-      // Simulate minute ticks
-      for (let tick = 0; tick < 5; tick++) {
-        updateCallback()
-      }
-
-      expect(updateCallback).toHaveBeenCalledTimes(5)
-
-      // Format should be consistent across updates
-      const format = '2h 45m'
-      expect(format).toMatch(/^\d+h\s\d+m$/)
-    })
-  })
-
-  describe('Delay Detection Integration', () => {
-    it('should identify delayed patients (> 3 hours)', () => {
-      const DELAY_THRESHOLD_MS = 10800000 // 3 hours
-      const elapsedTime = 11700000 // 3h 15m
-
-      const isDelayed = elapsedTime > DELAY_THRESHOLD_MS
-      expect(isDelayed).toBe(true)
-    })
-
-    it('should identify on-time patients (< 3 hours)', () => {
-      const DELAY_THRESHOLD_MS = 10800000
-      const elapsedTime = 9900000 // 2h 45m
-
-      const isDelayed = elapsedTime > DELAY_THRESHOLD_MS
-      expect(isDelayed).toBe(false)
-    })
-
-    it('should apply red color to delayed patients', () => {
-      const isDelayed = true
-      const textColor = isDelayed ? 'text-red-400' : 'text-zinc-300'
-      expect(textColor).toBe('text-red-400')
-    })
-
-    it('should apply normal color to on-time patients', () => {
-      const isDelayed = false
-      const textColor = isDelayed ? 'text-red-400' : 'text-zinc-300'
-      expect(textColor).toBe('text-zinc-300')
-    })
-  })
-
-  describe('Edge Cases', () => {
-    it('should handle patient just admitted', () => {
-      const admissionTime = Date.now() - 10000 // 10 seconds ago
-      const elapsed = Date.now() - admissionTime
-
-      const totalMinutes = Math.floor(elapsed / 60000)
-      const formatted = totalMinutes < 1 ? '< 1m' : `${totalMinutes}m`
-
-      expect(formatted).toBe('< 1m')
-    })
-
-    it('should handle multi-day patients', () => {
-      const admissionTime = Date.now() - 172800000 // 48 hours ago
-      const elapsed = Date.now() - admissionTime
-
-      const totalMinutes = Math.floor(elapsed / 60000)
-      const hours = Math.floor(totalMinutes / 60)
-      const minutes = totalMinutes % 60
-
-      expect(hours).toBe(48)
-      expect(minutes).toBeLessThan(60)
-    })
-
-    it('should handle null/invalid admission time', () => {
-      const admissionTime = null
-      const formatted = admissionTime ? 'valid' : 'N/A'
-      expect(formatted).toBe('N/A')
-    })
-  })
-
-  describe('Performance Tests', () => {
-    it('should calculate format in < 1ms', () => {
-      const formatElapsedTime = (ms: number) => {
-        const totalMinutes = Math.floor(ms / 60000)
-        if (totalMinutes < 1) return '< 1m'
-        const hours = Math.floor(totalMinutes / 60)
-        const minutes = totalMinutes % 60
-        if (hours === 0) return `${minutes}m`
-        return minutes === 0 ? `${hours}h` : `${hours}h ${minutes}m`
-      }
-
-      const start = performance.now()
-      formatElapsedTime(9900000)
-      const end = performance.now()
-
-      expect(end - start).toBeLessThan(1)
-    })
-
-    it('should handle 12 beds updating every minute efficiently', () => {
-      const bedCount = 12
-      const updateCount = 60 // 1 hour of updates
-
-      const start = performance.now()
-      for (let i = 0; i < bedCount * updateCount; i++) {
-        const elapsed = 9900000
-        const totalMinutes = Math.floor(elapsed / 60000)
-        const hours = Math.floor(totalMinutes / 60)
-        const mins = totalMinutes % 60
-        // Simulate formatting
-        const formatted = `${hours}h ${mins}m`
-        expect(formatted).toBeDefined()
-      }
-      const end = performance.now()
-
-      expect(end - start).toBeLessThan(100) // Should be fast
-    })
-  })
 })
 
-// ─────────────────────────────────────────────────────────────────────────────
-// US-1.5 Acceptance Criteria — Filter and Sort Delayed Beds
-// ─────────────────────────────────────────────────────────────────────────────
-
-describe('US-1.5 Acceptance Criteria — Filter and Sort Delayed Beds', () => {
-  // Shared bed fixtures
-  const makeFilterBed = (id: string, isDelayed: boolean, elapsedMs: number | null) => ({
-    id,
-    isDelayed,
-    elapsedTimeMs: elapsedMs,
-  })
-
-  const BEDS = [
-    makeFilterBed('B1', true,  12_600_000),  // 3h 30m — delayed
-    makeFilterBed('B2', false,  2_700_000),  // 45m     — on time
-    makeFilterBed('B3', true,  18_000_000),  // 5h      — delayed (longest)
-    makeFilterBed('B4', false,          null), // empty
-    makeFilterBed('B5', true,  10_800_000),  // 3h      — delayed
-  ]
-
-  describe('AC-1: "Show Delayed Only" filter button', () => {
-    it('should isolate delayed beds when filter is active', () => {
-      const delayed = BEDS.filter((b) => b.isDelayed)
-      expect(delayed).toHaveLength(3)
-      expect(delayed.every((b) => b.isDelayed)).toBe(true)
-    })
-
-    it('should restore all beds when filter is turned off', () => {
-      // Apply then remove
-      let result = BEDS.filter((b) => b.isDelayed)
-      expect(result).toHaveLength(3)
-      result = BEDS // clear
-      expect(result).toHaveLength(5)
-    })
-
-    it('should correctly count delayed beds for the badge', () => {
-      const delayedCount = BEDS.filter((b) => b.isDelayed).length
-      expect(delayedCount).toBe(3)
-    })
-  })
-
-  describe('AC-2: Sort by delay duration', () => {
-    it('should sort all beds longest-wait-first', () => {
-      const sorted = [...BEDS].sort((a, b) => (b.elapsedTimeMs ?? 0) - (a.elapsedTimeMs ?? 0))
-      const ids = sorted.map((b) => b.id)
-      // B3(18M) → B1(12.6M) → B5(10.8M) → B2(2.7M) → B4(null→0)
-      expect(ids).toEqual(['B3', 'B1', 'B5', 'B2', 'B4'])
-    })
-
-    it('should sort only the delayed subset when filter is also active', () => {
-      const delayed = BEDS.filter((b) => b.isDelayed)
-      const sorted = [...delayed].sort((a, b) => (b.elapsedTimeMs ?? 0) - (a.elapsedTimeMs ?? 0))
-      expect(sorted.map((b) => b.id)).toEqual(['B3', 'B1', 'B5'])
-    })
-
-    it('should treat null elapsedTimeMs as 0 when sorting', () => {
-      const sorted = [...BEDS].sort((a, b) => (b.elapsedTimeMs ?? 0) - (a.elapsedTimeMs ?? 0))
-      expect(sorted[sorted.length - 1].id).toBe('B4') // null goes last
-    })
-  })
-
-  describe('AC-3: Filter state persists during session', () => {
-    it('should serialise filter state to a plain JSON object', () => {
-      const state = { showDelayedOnly: true, sortOrder: 'desc' }
-      const serialised = JSON.stringify(state)
-      const parsed = JSON.parse(serialised)
-      expect(parsed.showDelayedOnly).toBe(true)
-      expect(parsed.sortOrder).toBe('desc')
-    })
-
-    it('should survive a round-trip through sessionStorage', () => {
-      const SESSION_KEY = 'ewtcs:bedFilter'
-      sessionStorage.setItem(SESSION_KEY, JSON.stringify({ showDelayedOnly: true, sortOrder: 'desc' }))
-      const restored = JSON.parse(sessionStorage.getItem(SESSION_KEY) ?? '{}')
-      expect(restored.showDelayedOnly).toBe(true)
-      expect(restored.sortOrder).toBe('desc')
-      sessionStorage.removeItem(SESSION_KEY)
-    })
-
-    it('should fall back gracefully on corrupted storage', () => {
-      const raw = 'not-valid{{'
-      let parsed: { showDelayedOnly: boolean; sortOrder: string }
-      try {
-        parsed = JSON.parse(raw)
-      } catch {
-        parsed = { showDelayedOnly: false, sortOrder: 'none' }
-      }
-      expect(parsed.showDelayedOnly).toBe(false)
-      expect(parsed.sortOrder).toBe('none')
-    })
-  })
-
-  describe('AC-4: Clear indication when filter is active', () => {
-    it('isFilterActive should be true when showDelayedOnly is true', () => {
-      const isFilterActive = (showDelayedOnly: boolean, sortOrder: string) =>
-        showDelayedOnly || sortOrder !== 'none'
-
-      expect(isFilterActive(true, 'none')).toBe(true)
-    })
-
-    it('isFilterActive should be true when sort is active', () => {
-      const isFilterActive = (showDelayedOnly: boolean, sortOrder: string) =>
-        showDelayedOnly || sortOrder !== 'none'
-
-      expect(isFilterActive(false, 'desc')).toBe(true)
-    })
-
-    it('isFilterActive should be false when nothing is active', () => {
-      const isFilterActive = (showDelayedOnly: boolean, sortOrder: string) =>
-        showDelayedOnly || sortOrder !== 'none'
-
-      expect(isFilterActive(false, 'none')).toBe(false)
-    })
-  })
-
-  describe('AC-5: One-click to clear filter', () => {
-    it('should reset both filter and sort to defaults', () => {
-      // Simulate state: both active → both cleared
-      let showDelayedOnly = true
-      let sortOrder = 'desc'
-
-      // clearFilter
-      showDelayedOnly = false
-      sortOrder = 'none'
-
-      expect(showDelayedOnly).toBe(false)
-      expect(sortOrder).toBe('none')
-    })
-
-    it('isFilterActive should be false immediately after clear', () => {
-      const isFilterActive = (f: boolean, s: string) => f || s !== 'none'
-      expect(isFilterActive(false, 'none')).toBe(false)
-    })
-  })
-})
+// Full AC suite is split into focused files:
+// - AC-US1.3-time-display.test.ts     (AC-1: time formatting)
+// - AC-US1.3-time-advanced.test.ts    (AC-3/4/5 + cross-AC + performance)
+// - AC-US1.3-time-validation.test.ts  (AC-2: timer, edge cases, delay)
+// - AC-US1.5.integration.test.ts      (US-1.5: filter and sort)
