+++ src/features/bed-dashboard/components/BedCard.tsx
@@ -1,7 +1,7 @@
 // Bed Card Component
 // Epic 1: Nurse Desk Bed Dashboard
 
-import { memo, type MouseEvent, Fragment } from 'react'
+import { memo, type MouseEvent } from 'react'
 import { Card, CardContent } from '@/shared/components/ui/card'
 import { Clock, AlertTriangle, Hourglass } from 'lucide-react'
 import type { BedWithElapsedTime, DispositionDelayReason } from '../types/bed'
@@ -9,25 +9,7 @@ import { DISPOSITION_DELAY_REASON_LABELS } from '../types/bed'
 import { formatElapsedTime, getStageColorClasses, getDelayColorClasses } from '../lib/utils'
 import { useElapsedTime } from '../hooks/useElapsedTime'
 import { cn } from '@/shared/lib/utils'
-
-function highlightMatch(text: string, query?: string) {
-  if (!query) return text
-  const q = query.trim().toLowerCase()
-  if (!q) return text
-  const lower = text.toLowerCase()
-  const idx = lower.indexOf(q)
-  if (idx === -1) return text
-  const before = text.slice(0, idx)
-  const match = text.slice(idx, idx + q.length)
-  const after = text.slice(idx + q.length)
-  return (
-    <Fragment>
-      {before}
-      <span className="bg-yellow-300 text-black px-1 rounded">{match}</span>
-      {after}
-    </Fragment>
-  )
-}
+import { highlightMatch } from '../lib/highlight-match'
 
 const REASON_OPTIONS = Object.entries(DISPOSITION_DELAY_REASON_LABELS) as [
   DispositionDelayReason,
